Code refactoring

+++ service/src/main/java/greencity/service/EcoNewsServiceImpl.java
@@ -317,12 +317,16 @@ public class EcoNewsServiceImpl implements EcoNewsService {
     @Override
     public void delete(Long id, UserVO user) {
         EcoNewsVO ecoNewsVO = findById(id);
+        if (user.getRole() != Role.ROLE_ADMIN && user.getRole() != Role.ROLE_USER) {
+            throw new BadRequestException(ErrorMessage.USER_HAS_NO_PERMISSION);
+        }
         if (user.getRole() != Role.ROLE_ADMIN && !user.getId().equals(ecoNewsVO.getAuthor().getId())) {
             throw new AccessDeniedException(ErrorMessage.USER_HAS_NO_PERMISSION);
         }
         String accessToken = httpServletRequest.getHeader(AUTHORIZATION);
-        CompletableFuture.runAsync(
-            () -> ratingCalculation.ratingCalculation(RatingCalculationEnum.DELETE_ECO_NEWS, user, accessToken));
+        CompletableFuture.runAsync(() ->
+                ratingCalculation.ratingCalculation(RatingCalculationEnum.DELETE_ECO_NEWS, user, accessToken)
+        );
         ecoNewsRepo.deleteById(ecoNewsVO.getId());
     }
